feat: LinkAsButtonコンポーネントを新規作成し、内部リンクと外部リンクの処理を統合
