+++ src/application/web/routes/admin._layout.tsx
@@ -1,7 +1,7 @@
 import { useEffect } from 'react'
 import { Outlet, useNavigate } from 'react-router'
 import { toast } from 'sonner'
-import LinkAsButton from '../components/linkAsButton'
+import { LinkAsButton } from '../components/link'
 import getApiClientForClient from '../infrastructure/api'
 
 export default function AdminLayout() {
